心拍バースト送信機能 double型で送信していて変な値になっていたのを修正

+++ BluetoothSPP.ino
@@ -48,7 +48,7 @@ void sendBarstBeatAvg() {
     while (str[i] != '\0') {
       i++;
     }
-    sprintf(&str[i], "%d,", display_queue(b));
+    sprintf(&str[i], "%d,", (int)display_queue(b));
     b++;
   }
   sendSppMsg(str);
